backend/requirements.txt
Django==5.2.8
djangorestframework
djangorestframework-simplejwt
python-dotenv
psycopg2-binary
PyJWT
requests
django-cors-headers
djangorestframework-camel-case
